apps/ocr-fallback/requirements-vl.txt
torch>=2.0.0
transformers>=5.0.0
accelerate>=0.20.0
Pillow>=9.0.0
sentencepiece
